Separated Posts and Courses parse logic
