+++ utils/getFeed.js
@@ -1,3 +1,6 @@
+/* Internal dependencies */
+import { parseFeedItem as parsePostFeedItem } from "./getPostFeed";
+
 /**
  * @typedef  {Object}     Feed
  * @property {string}     title       The title of the website.
@@ -26,7 +29,7 @@
  *
  * @returns {string|undefined} The string result of the xpath query.
  */
-function getXPathText( xpath, document, context = null, nsResolver = null ) {
+export function getXPathText( xpath, document, context = null, nsResolver = null ) {
 	const tagname = xpath.replace( "child::", "" );
 	const elementExists = document.getElementsByTagName( tagname );
 
@@ -62,31 +65,6 @@ function getFeedMeta( parsed ) {
 }
 
 /**
- * Returns a single feed item from a snapshot.
- *
- * @param {Document}        parsed     A parsed XML document.
- * @param {Node}            snapshot   A snapshot returned from the snapshotItem method of a XPathResult.
- * @param {XPathNSResolver} nsResolver A namespace resolver for the parsed document.
- *
- * @returns {FeedItem} The FeedItem representing the provided snapshot.
- */
-function getFeedItem( parsed, snapshot, nsResolver ) {
-	const item = {};
-
-	item.title       = getXPathText( "child::title", parsed, snapshot );
-	item.link        = getXPathText( "child::link", parsed, snapshot );
-	item.content     = getXPathText( "child::content:encoded", parsed, snapshot, nsResolver );
-	item.description = getXPathText( "child::description", parsed, snapshot );
-	item.creator     = getXPathText( "child::dc:creator", parsed, snapshot, nsResolver );
-	item.date        = getXPathText( "child::pubDate", parsed, snapshot );
-	item.image       = getXPathText( "child::content:image", parsed, snapshot, nsResolver );
-	item.shopUrl     = getXPathText( "child::content:shop_url", parsed, snapshot, nsResolver );
-	item.id          = getXPathText( "child::content:slug", parsed, snapshot, nsResolver );
-
-	return item;
-}
-
-/**
  * Returns a ORDERED_NODE_SNAPSHOT_TYPE XpathResult for the given xpath query on the provided document.
  *
  * @param {string}          xpath      The xpath query to run.
@@ -109,9 +87,9 @@ function getXPathSnapshots( xpath, document, context = null, nsResolver = null )
  *
  * @returns {FeedItem[]} An array of FeedItem objects.
  */
-function getFeedItems( parsed, nsResolver, maxItems ) {
+function getFeedItems( parsed, nsResolver, maxItems, parseFeedItem ) {
 	const snapshots = getXPathSnapshots( "/rss/channel/item", parsed );
-	let count     = snapshots.snapshotLength;
+	let count       = snapshots.snapshotLength;
 	const items     = [];
 
 	if ( maxItems !== 0 ) {
@@ -120,7 +98,7 @@ function getFeedItems( parsed, nsResolver, maxItems ) {
 
 	for ( let i = 0; i < count; i++ ) {
 		const snapshot = snapshots.snapshotItem( i );
-		items.push( getFeedItem( parsed, snapshot, nsResolver ) );
+		items.push( parseFeedItem( parsed, snapshot, nsResolver ) );
 	}
 
 	return items;
@@ -134,15 +112,15 @@ function getFeedItems( parsed, nsResolver, maxItems ) {
  *
  * @returns {Promise.<Feed>} A promise which resolves with the parsed Feed.
  */
-export function parseFeed( raw, maxItems = 0 ) {
+export function parseFeed( raw, maxItems = 0, parseFeedItem ) {
 	return new Promise( function( resolve, reject ) {
 		try {
 			const parser     = new DOMParser();
 			const parsed     = parser.parseFromString( raw, "application/xml" );
 			const nsResolver = parsed.createNSResolver( parsed.documentElement );
 
-			const result   = getFeedMeta( parsed );
-			result.items = getFeedItems( parsed, nsResolver, maxItems );
+			const result = getFeedMeta( parsed );
+			result.items = getFeedItems( parsed, nsResolver, maxItems, parseFeedItem );
 
 			resolve( result );
 		} catch ( error ) {
@@ -154,15 +132,17 @@ export function parseFeed( raw, maxItems = 0 ) {
 /**
  * Grabs an RSS feed from the requested URL and parses it.
  *
- * @param {string} url      The URL the feed is located at.
- * @param {int}    maxItems The amount of items you wish returned, 0 for all items.
+ * @param {string}   url           The URL the feed is located at.
+ * @param {int}      maxItems      The amount of items you wish returned, 0 for all items.
+ * @param {function} parseFeedItem Function used to parse individual feed items, parses post feed items
+ *                                 by default.
  *
  * @returns {Promise.<Feed>} The retrieved feed.
  */
-export default function getFeed( url, maxItems = 0 ) {
+export default function getFeed( url, maxItems = 0, parseFeedItem = parsePostFeedItem ) {
 	return fetch( url ).then( function( response ) {
 		return response.text();
 	} ).then( function( raw ) {
-		return parseFeed( raw, maxItems );
+		return parseFeed( raw, maxItems, parseFeedItem );
 	} );
 }
